Refactor ios code
