+++ Fruitties/iosApp/iosApp/ui/ContentView.swift
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
+import Foundation
 import SwiftUI
 import shared
-import Foundation
 
 struct ContentView: View {
     /// Injects the `IOSViewModelStoreOwner` from the environment, which manages the lifecycle of `ViewModel` instances.
@@ -37,7 +37,7 @@ struct ContentView: View {
             VStack {
                 Text("Fruitties").font(.largeTitle).fontWeight(.bold)
                 NavigationLink {
-                    ViewModelStoreOwnerProvider{
+                    ViewModelStoreOwnerProvider {
                         CartView()
                     }
                 } label: {
@@ -49,12 +49,18 @@ struct ContentView: View {
                 Observing(mainViewModel.homeUiState) { homeUIState in
                     ScrollView {
                         LazyVStack {
-                            ForEach(homeUIState.fruitties, id: \.self) { value in
-                                FruittieView(fruittie: value, addToCart: { fruittie in
-                                    Task {
-                                        mainViewModel.addItemToCart(fruittie: fruittie)
+                            ForEach(homeUIState.fruitties, id: \.self) {
+                                value in
+                                FruittieView(
+                                    fruittie: value,
+                                    addToCart: { fruittie in
+                                        Task {
+                                            mainViewModel.addItemToCart(
+                                                fruittie: fruittie
+                                            )
+                                        }
                                     }
-                                })
+                                )
                             }
                         }
                     }
@@ -70,7 +76,9 @@ struct FruittieView: View {
     var body: some View {
         HStack(alignment: .firstTextBaseline) {
             ZStack {
-                RoundedRectangle(cornerRadius: 15).fill(Color(red: 0.8, green: 0.8, blue: 1.0))
+                RoundedRectangle(cornerRadius: 15).fill(
+                    Color(red: 0.8, green: 0.8, blue: 1.0)
+                )
                 VStack {
                     Text("\(fruittie.name)")
                         .fontWeight(.bold)
@@ -79,9 +87,12 @@ struct FruittieView: View {
                         .frame(maxWidth: .infinity, alignment: .leading)
                 }.padding()
                 Spacer()
-                Button(action: { addToCart(fruittie) }, label: {
-                    Text("Add")
-                }).padding().frame(maxWidth: .infinity, alignment: .trailing)
+                Button(
+                    action: { addToCart(fruittie) },
+                    label: {
+                        Text("Add")
+                    }
+                ).padding().frame(maxWidth: .infinity, alignment: .trailing)
             }.padding([.leading, .trailing])
         }
     }
